MovieData.cs update
Also updated Movie Modle for attribute Revenue to be Int64 instead of Instead of Int32 because Avatar's Revenue was too high. lol

+++ LonghornCinemaFinalProject/LonghornCinemaFinalProject/Models/Movie.cs
@@ -27,7 +27,7 @@ namespace LonghornCinemaFinalProject.Models
         public DateTime ReleaseDate { get; set; }
 
         // Revenue
-        public Int32 Revenue { get; set; }
+        public Int64 Revenue { get; set; }
 
         // Runtime (in minutes)
         public Int32 Runtime { get; set; }
